Fix execution time of startup script
Before, `think-startup` was started by `/etc/rc.local`, which caused it to execute before the user logged in. As a result, X-related commands, such as disabling the touch screen through `xinput`, did not work. Now, `think-startup` is called as X starts (through a script placed in `/etc/X11/Xsession.d/`), so it works properly.