Move keycode mappings to udev hwdb config file
This removes the need for the think-keycodes system startup script,
which makes the mappings work with other init systems by default, such
as systemd. The 90-X220T-keyboard.hwdb file added in this commit is
modeled after the Lenovo X200 Tablet mappings that are part of upstream
udev.

Only the bindings necessary for the X220T are added in this commit. The
X200T config is already setup correctly upstream in udev, but if you
use an even older Thinkpad tablet, we may need to add support for it.

Note the following about the two keycode mappings:

- The screen rotation bezel key is still mapped to keycode 120
  ("scale") instead of the X200T's default of keycode 153 ("direction")
  because keycode 120 is already mapped to an appropriate keysym in X.
  Keycode 153 is not, so the user would have to use a utility such as
  xmodmap to map the keycode to a keysym for use in X.

- The circular arrow bezel key is now mapped to keycode 154
  ("cyclewindows") to match the X200T. Before, it was mapped to nothing.