Stop updating udev hwdb when packaging
The `udevadm hwdb --update` line needs to be run as root upon
installation to update udev's hardware database with the new information
from 90-X220T-keyboard.hwdb. We should run this when installing to
root (i.e. DESTDIR is empty), but not when DESTDIR is anything
else (e.g. when packaging think-rotate).