Fix qdbus on Arch Linux systems without qtchooser
Arch Linux installs symlinks in /usr/bin to the Qt binaries with names
in the form `name-qt4` for Qt 4, `name-qt5` for Qt 5, etc., so with just
Qt 4 installed, `qdbus-qt4` is on the system path but not `qdbus`.  The
`qtchooser` package is necessary to add `qdbus` to the system path.
This commit makes qdbus work properly even without `qtchooser` installed
because it calls `qdbus-qt4` when available and `qdbus` otherwise.