Fix think-rotate-acpi-hook
For each hook file in /etc/acpi/events, `acpid` registers only the last
event listed. As a result, the handler for the `video/tabletmode...`
event was never called. Splitting `think-rotate-acpi-hook` into two
separate files fixes the issue.

Additionally, removed the execute permission from the ACPI event hooks
because it is not needed.