Change table structure to match others
While I was working on the Debian package, I looked up the Debian
package names in the second column of the dependencies. However, when I
did the optional dependencies, I used the second column, which is the
program name.

Therefore, I matched the column order of the last table with the other
two tables.