Add support for changing subpixel order under XFCE
This also fixes the case (as on XFCE) where gsettings is present but
the org.gnome.settings-daemon... schema is not. Before this commit,
calling gsettings on XFCE with the nonexistent schema was failing and
causing thinkpad-rotate to terminate early.

Note that xfconf-query works properly when thinkpad-rotate is called
manually by the user but does not seem to work when thinkpad-rotate is
called from thinkpad-rotate-hook.