Also match more characters with acpid script
For some reason, my X220 Tablet would show this with `acpi_listen`:

    TBLT 0000008A 000000001 K

The manual for `acpid` says that the `event=` line contains a regular
expression that is “matched” against the event. Therefore, it did not
match this event, skipping the rotation.

After installing and restarting acpid, the `K` in the event was gone. I
observed that the hardware triggered rotation does not work at some
points, restarting the laptop makes them work again. I assume that the
`K` thing was the issue here.