Add network.connection_name option
I have a “DHCP” and a “Wired Shared” connection in network manager which
are both wired. For some reason, nmcli would give me “Wired Shared” as
the first one, restarting that one. That broke the restarting feature
for me, so I added this new option that lets me fix the name to “DHCP”
without removing the other one from network manager.