Make selection of Ethernet connection predictable
Before, the first Ethernet connection in the list given by
`nmcli con show` was selected. However, this was unpredictable and
different connections could be selected on each reboot. Now, the
selected connection is the lexicographically first Ethernet connection.

Another option would be to select the Ethernet connection based on the
physical device name (which could be obtained via the MAC address). See
issue #75. However, this adds needless complication. The user can
manually specify the one they want in their configuration file if the
wrong one is selected by default.