Detect single logged-in user as the graphical user
The current heuristic (searching the output of `who -u` for an X
display) doesn't always work. For example, on my computer, the output
of `who -u` is simply:

    jim      tty1         2017-02-24 12:08 02:21         730

This commit changes `get_graphicsl_user()` so that if `who -u` lists a
single logged-in user, that user is detected as the graphical user.